refactor: delete unuseful imports, components and data, change folders structure, update imports
