--- src/components/ChatMessage/ChatMessage.scss
@@ -1,27 +1,0 @@
-@import '../../styles/variables.scss';
-.chat-message {
-    width: 50%;
-    height: auto;
-    background-color: $color-text-light;
-    border-radius: 10px;
-    padding: 1rem;
-    font-weight: 500;
-    font-size: 12px;
-    line-height: 15px;
-    color: $color-black;
-    margin: 1rem 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    &_mate {
-        align-self: flex-start;
-    }
-    &_owner {
-        background-color: $color-main-second;
-        align-self: flex-end;
-    }
-    &__time {
-        font-size: 1rem;
-        align-self: flex-end;
-    }
-}
