--- src/data/chats.ts
@@ -1,16 +1,0 @@
-import { ChatItemPreview } from 'components/ChatItem/ChatItem';
-
-export const CHATS: ChatItemPreview[] = [
-  {
-    name: 'Vlad',
-    message: 'И Human Interface Guidelines и Material Design рекомендуют...',
-    time: '16:30',
-    unread: '2',
-  },
-  {
-    name: 'Илья',
-    message: 'Друзья, у меня для вас особенный выпуск новостей!...',
-    time: '15:12',
-    unread: '4',
-  },
-];
