--- src/data/userData.ts
@@ -1,29 +1,0 @@
-export const userData = [
-  {
-    title: 'Email',
-    data: 'mail@pochta.ru',
-    type: 'email',
-  },
-  {
-    title: 'Login',
-    data: 'ivan',
-    type: 'text',
-  },
-
-  {
-    title: 'FirstName',
-    data: 'Ivan',
-    type: 'text',
-  },
-  {
-    title: 'SecondName',
-    data: 'Ivanov',
-    type: 'text',
-  },
-
-  {
-    title: 'Phone',
-    data: '+7 000 000 00 00',
-    type: 'tel',
-  },
-];
