--- src/data/userPassword.ts
@@ -1,7 +1,0 @@
-export const userData = [
-  {
-    title: 'Password',
-    data: 'Password8Uhgfjhf',
-    type: 'password',
-  },
-];
