+++ src/services/startApp.ts
@@ -1,10 +1,10 @@
 import { Dispatch } from 'store/Store';
 import AuthAPI from 'API/AuthAPI';
-import { isApiReturnedError } from 'utils/isApiReturnedError';
-import { transformUserObject } from 'utils/transformUserObject';
+import { isApiReturnedError } from 'utils/checkers and validators/isApiReturnedError';
+import { transformUserObject } from 'utils/transformers/transformUserObject';
 import { ChatFromServer, UserFromServer } from 'API/typesAPI';
 import ChatsAPI from 'API/ChatsAPI';
-import { transformChatsObject } from 'utils/transformChatsObject';
+import { transformChatsObject } from 'utils/transformers/transformChatsObject';
 import { getAvatar } from './userData';
 
 const authApi = new AuthAPI();
