+++ src/utils/checkers and validators/isApiReturnedError.ts
