+++ src/utils/checkers and validators/isObject.ts
