+++ src/utils/checkers and validators/isStringEqual.ts
