+++ src/utils/cloneDeep.ts
@@ -1,4 +1,4 @@
-import { isObject } from './isObject';
+import { isObject } from './checkers and validators/isObject';
 
 export function cloneDeep<T extends object = object>(obj: T) {
   const clone = (Array.isArray(obj) ? [] : {}) as T;
