+++ src/utils/set.ts
@@ -1,6 +1,6 @@
 import deepMerge from './deepMerge';
 import getObjectFromPath from './getObjectFromPath';
-import { isObject } from './isObject';
+import { isObject } from './checkers and validators/isObject';
 
 function set(object: Indexed | unknown, path: string, value: unknown): Indexed | unknown {
   if (typeof path !== 'string') {
